Convenience function for formatting to a tree
